requirements.txt
fastapi
uvicorn
python-dotenv
requests
PyJWT
cryptography
sqlalchemy
alembic
python-jose[cryptography]